% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger.getLevelNumber}
\alias{logger.getLevelNumber}
\title{Returns numeric value for string}
\usage{
logger.getLevelNumber(level)
}
\arguments{
\item{level}{string representation of logger message level}
}
\value{
integer corresponding to the message level
}
\description{
Returns numeric value for string
}
\details{
ALL   =  0
DEBUG = 10
INFO  = 20
WARN  = 30
ERROR = 40
ALL   = 99
}
\author{
Rob Kooper
}
